got save and load to work
